Do not test for sb-eval feature
SBCL 2.0.1 removed :sb-eval from *features* (and added it to an internal
features list). Additionally the #+ reader macro now produces a full warning if
a feature present on the internal features list is tested for.

This commit tests for the presence of sb-kernel:lexical-environment-too-complex
directly instead of looking at the sb-eval feature.