Ignore spurious errors in SBCL compilation.
Ignore errors in sb-grovel-unknown-constant-condition-p.

An error is raised if `simple-condition-format-control` is a formatter
functionand is is passed to `string`. This now mirrors
`match-condition-p` which also ignores errors when comparing a simple
condition to a string.

Test added to test-utilities.script for this issue.

Thanks to Eric Timmons for the fix.