Fix bug which causes compilation to leave empty temporary files on Genera
Genera uses versioned pathnames. TMPIZE-PATHNAME creates a temporary file
to ensure a unique name. If we leave that file, Genera will create a new
version (i.e., version 2) when its caller opens the file for output. After
that file is renamed, the original temporary file remains. Since Genera runs
on a single core processor, the possibility of a name collision is minimal
if not nil. So, don't keep the temporary file.