Pin ABCL to 1.7.1 on JDK 8 for testing
ABCL 1.8.0 has an issue where the PID of a process cannot be retrieved, causing
a failure in test-run-program-unix.script. This has been reported to and
reproduced by ABCL devs.

Additionally, ABCL 1.8.0 fails test-around-compile.script. I'm still trying to
figure out if this is ABCL's fault or ASDF's fault, but it's likely ABCL's
given that it works in 1.7.1.

Pin to JDK 8 because newer JDKs have issues even launching external
processes (that issue has been fixed in 1.8.0).