Don't warn about ASCII encodings (subset of UTF-8)
Currently warnings are thrown by concatenate-source-op every time ASDF
is loaded if UTF-8 is the default encoding because the ASDF source is
ASCII which is not equal to UTF-8.  This commit adds a special case to
the encoding equality check so that ASCII is recognized as a subset of
UTF-8.