Fix Genera's special case to skip /usr/share/common-lisp
OS-COND evaluates its clauses at compile time and only includes the relevant clause.
Change the clause in XDG-DATA-DIRS for Genera to check whether we're on a VLM running
on macOS at runtime. Otherwise, the binary we distribute with Genera will either
include /usr/share/common-lisp or it won't depending on which machine I use to compile.