rewrite docs makefile for less spurious error-looking output
I got pissed off that the text "failed. See log in <rule>.log"
appeared in successful doc makefile executions, so I made that no
longer the case. Now, failure messages print when a command fails, but
will not appear at all in the output of a successful run.

Also, processes' error output was not being redirected to logfiles,
because the `2>&1` was before the `> logfile` and shell redirects are
stupid. Now both stdout and stderr of subprocesses are redirected to
logfiles.