Add extension points for component parsing.
Eric Timmons writes [edited for compression]:

I decided to make a system class that infers components instead of systems called [asdf-inferred-components](https://gitlab.common-lisp.net/etimmons/asdf-inferred-components).

I found myself wanting a couple more points to intercede in the parsing of systems. This merge adds them.

At the moment they are not exported from the ASDF package, but we anticipate they will be exported after using them in practice for a while.

`class-for-type` is now a generic function that allows a custom `parent-component` class to control the mapping of component types (`:file`, `:module`, etc.) to classes.

I also turned `compute-component-children` into a GF (no backward compatibility issues there since the `defun` version of it has yet to be released) and moved a bit of the work it did back into `parse-component-form`.