Use REGISTERED-SYSTEM to find preloaded system instead of FIND-SYSTEM
FIND-SYSTEM may cause the system to be loaded from the file
system. REGISTERED-SYSTEM won't load it from the file system, but will return
something from the file system if it's already been loaded from there.

It's safe to do this as the preloaded systems are always considered to be
registered.