Fix test-mutual-redefinition.script
Since 3.3.2 / 069cd2a66bc64a38d87b083aca2cbddeee0127ed, the primary system name
of a misnamed system is derived from the pathname of the .asd file that defines
it, not the name of the system itself. This means that within the same cache
session, any operation on a misnamed system will depend on DEFINE-OP for the
file where it was already found and will *not* depend on a DEFINE-OP for the
"syntactic" primary name.

The test test-mutual-redefinition.script has been broken since (at least)
then. However, it was missed due to UPGRADE-ASDF always returning T. This
resulted in every load-system in this test running with an isolated cache.