Better error message for import-from in define-package.
Resolve "UIOP:DEFINE-PACKAGE signals a TYPE-ERROR when trying to IMPORT-FROM a non-existent package" (#98)

The code handling other options to `define-package` used `find-package*`, 
which (by default) errors if it can't find the package it's looking for.  The
exception was the code for `:import-from`. This caused cryptic error messages
(issue #98) when trying to import from a package that did not exist.

Switched to using `find-package*`, added a new exception class, and added a test.