Fix test-module-excessive-depend.script
This test has probably been broken for a very long time, but it was missed due
to UPGRADE-ASDF returning T all the time. This meant that the first LOAD-OP was
performed in a new session, *not* the session created at the top of the
file. Once UPGRADE-ASDF was fixed, then the done status of the LOAD-OP was
persisted, resulting in the next LOAD-OP being a no op.

However, this breakage wasn't immediately apparent because FILE-WRITE-DATE was
compared with #'>= and the PLAN was not asserted to actually contain the
operations we wanted to see. So on a fast computer (unless you got really
unlucky) you'd never see the failed test.