Better handling of missing packages in DEFINE-PACKAGE :IMPORT-FROM
Signal a style warning if we `:import-from` using a package that does not exist and an empty symbol list. Remains an error to `:import-from` a package that does not exist with a non-empty symbol list.

Add some tests to try and ensure `:uiop/package` and `:uiop/package-local-nicknames` are not modified.