Test for issue 124.
The issue involves loading code that does not have an initial package specification. In that case, if `*package*` is bound to a package that does not import `defpackage`, a file that defines a package and then switches to it, may fail.

This test demonstrates the failure mode, so that it can be fixed later.